适配新的图片添加策略

+++ message_service_1/src/main/java/org/xzx/bean/messageBean/Message.java
@@ -22,7 +22,7 @@ public class Message {
 
     private String sub_type;
 
-    private int message_id;
+    private long message_id;
 
     private long user_id;
