+++ message_service_1/src/main/java/org/xzx/bean/messageUtil/MessageCounter.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class MessageCounter {
-    private int group_id;
+    private long group_id;
 
     private int messageCount;
