+++ message_service_1/src/main/java/org/xzx/service/Gocq_service.java
@@ -23,17 +23,17 @@ public class Gocq_service {
     @Value("${gocq.basicUrl}")
     private String basicUrl;
 
-    public void send_group_message(int group_id, String message) {
+    public void send_group_message(long group_id, String message) {
         SendGroupMessage groupMessage = new SendGroupMessage(group_id, message);
         System.out.println(springRestService.postWithObject(basicUrl + "send_group_msg", groupMessage, String.class));
     }
 
-    public void send_private_message(int user_id, String message) {
+    public void send_private_message(long user_id, String message) {
         PrivateMessage privateMessage = new PrivateMessage(user_id, message);
         System.out.println(springRestService.postWithObject(basicUrl + "send_private_msg", privateMessage, String.class));
     }
 
-    public String get_message(int message_id) {
+    public String get_message(long message_id) {
         JsonNode jsonNode = springRestService.getForObject(basicUrl + "get_msg?message_id=" + message_id, JsonNode.class);
         return jsonNode.get("data").get("message").asText();
     }
